requirements.txt
beautifulsoup4==4.11.1
humanfriendly==10.0
mutagen==1.46.0
progressbar33==2.4
requests==2.28.1
